+++ core/auth/signature.go
@@ -31,8 +31,8 @@ func (a *signatureAuthImpl) verify(credential *model.Credential, verification st
 	return nil, errors.New(logutils.Unimplemented)
 }
 
-func (a *signatureAuthImpl) checkCredentials(accountAuthType model.AccountAuthType, creds string, l *logs.Log) (string, *bool, error) {
-	return "", nil, nil
+func (a *signatureAuthImpl) checkCredentials(accountAuthType model.AccountAuthType, creds string, l *logs.Log) (string, error) {
+	return "", nil
 }
 
 //initSignatureAuth initializes and registers a new signature auth instance